qemux86: system.conf: replace statusfile by new data-directory
This makes use of the new 'data-directory' introduced by RAUC 1.8 which
will be used for adaptive updates and further future features.

The data-directory implies a default storage location for the
statusfile, too. Thus we can safely remove this.

Note: The RAUC 1.8 release is provided by meta-rauc's 'kirkstone' or
'master' branch.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>