qemux86: core-image-minimal: align image to 4K
This sets IMAGE_ROOTFS_ALIGNMENT which will align the generated root
file system size to 4K.
The 4K alignment is a precondition to for using adaptive method
'block-hash-index' which operates on 4K blocks.

Additionally, we use the -i and -b parameters of mkfs.ext4 to ensure
that the ext4 inode and block size is also 4k-aligend.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>