meta-rauc-nxp: Adding u-boot configuration
Adding userspace configuration for accessing u-boot environment.
Adding boot script.

The boot script will attempt to boot from both slots 3 times.
Attemting to boot means running the boot detection command used in the
default boot sequence for cubox-i. It will run u-boot script or
kernel using syslinux config found on the partition.

The boot script will save the current environment, effectively creating
it if not existing. The u-boot environment is stored in raw format on
the SD card and some flashing mechanisms (bmaptool) do NOT clean it.
This can lead to the environment being preserved after flashing an
image to the SD card, possibly cause issues.

The simplest way to avoid this is to clean it manually using dd or
similar tool before flashing the image. Example:
    dd if=/dev/zero of=/dev/sdX count=10000

Signed-off-by: Atanas Bunchev <atanas.bunchev@konsulko.com>