meta-rauc-tegra: switch to using 'verity' bundle format
The kernel configuration already enabled 'verity' support, but the
bundle did not.

The 'verity' format is the successor of the legacy 'plain' format which
is still supported by RAUC but should not be used for new projects.

Also note that only the 'verity' format allows using RAUC features as
streaming, encryption and adaptive updates.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>