treewide: remove pre-generated certificates
For software updates, security is quite crucial and using pre-generated
standard keys is not something that people should get used to.
We have a dedicated script to generate custom certificates:

  ./create-example-keys.sh

Thus this commit removes all pre-generated certificates together with
the RAUC_KEY_FILE and RAUC_CERT_FILE definitions in bundle recipes and
the SRC_URI modifications in rauc bbappends.

Instead add the README.rst section on generating keys we have for
meta-rauc-qemux86 to all platforms.

Signed-off-by: Enrico Jörns <ejo@pengutronix.de>