meta-rauc-beagleplay: initial support for BeaglePlay
This commit adds support for the BeaglePlay, based on the TI AM625 SoC:
https://www.beagleboard.org/boards/beagleplay

The image is built for the eMMC storage available on the board. For
convenience, a flasher image is also available, which can be booted
from the SD card and flashes the initial image to the eMMC. After the
initial image is flashed, the OTA updates enable all subsequent
updates of the system.

Signed-off-by: António Oliveira <antonio.oliveira@konsulko.com>