.github/workflows: generate workflows from template
We want to build more layers in CI while avoiding too much redundancy in
the workflow definitions. Although this could be done using reusable
workflows or custom actions, templating them using jinja is much simpler
and more flexible.

To avoid conflicts with the GitHub Actions syntax, we use « and »
instead of { and }.

The only change to the existing workflows is the order in which
'bitbake-layers add-layer' is run.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>